+++ sdk/src/math/trade.ts
@@ -4,7 +4,7 @@ import {
 	PositionDirection,
 	SpotMarketAccount,
 } from '../types';
-import { BN } from '@project-serum/anchor';
+import { BN } from '@coral-xyz/anchor';
 import { assert } from '../assert/assert';
 import {
 	PRICE_PRECISION,