+++ tests/pyth.ts
@@ -1,4 +1,4 @@
-import * as anchor from '@project-serum/anchor';
+import * as anchor from '@coral-xyz/anchor';
 import { assert } from 'chai';
 
 import { BASE_PRECISION, BN, BulkAccountLoader } from '../sdk';
@@ -13,7 +13,7 @@ import {
 	initUserAccounts,
 	sleep,
 } from './testHelpers';
-import { Program } from '@project-serum/anchor';
+import { Program } from '@coral-xyz/anchor';
 
 import { Keypair, PublicKey } from '@solana/web3.js';